refactor #6. 공통 로직 global 패키지로 이동
