+++ backend/user/src/main/java/com/scf/user/global/error/BasicExceptionHandler.java
@@ -1,6 +1,5 @@
-package com.scf.user.member.global.error;
+package com.scf.user.global.error;
 
-import com.scf.user.member.global.error.exception.BusinessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
