+++ backend/user/src/main/java/com/scf/user/global/error/ErrorCode.java
@@ -1,0 +1,11 @@
+package com.scf.user.global.error;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+
+    String getMessage();
+
+    HttpStatus getHttpStatus();
+
+}
