--- backend/user/src/main/java/com/scf/user/profile/global/exception/ProblemNotFoundException.java
@@ -1,8 +1,0 @@
-package com.scf.user.profile.global.exception;
-
-public class ProblemNotFoundException extends RuntimeException {
-
-    public ProblemNotFoundException(String message) {
-        super(message);
-    }
-}
